requirements.txt
fast-flights>=3.0rc0
httpx
python-dotenv>=1.0.0
